+++ src/sentry/models/avatars/__init__.py
@@ -1,11 +1,9 @@
 from .base import AvatarBase
 from .control_base import ControlAvatarBase
 from .organization_avatar import OrganizationAvatar
-from .sentry_app_avatar import SentryAppAvatar
 
 __all__ = (
     "AvatarBase",
     "ControlAvatarBase",
     "OrganizationAvatar",
-    "SentryAppAvatar",
 )
