+++ src/sentry/sentry_apps/models/__init__.py
@@ -1,5 +1,6 @@
 from .platformexternalissue import PlatformExternalIssue
 from .sentry_app import SentryApp
+from .sentry_app_avatar import SentryAppAvatar
 from .sentry_app_component import SentryAppComponent
 from .sentry_app_installation import SentryAppInstallation
 from .sentry_app_installation_for_provider import SentryAppInstallationForProvider
@@ -14,4 +15,5 @@ __all__ = (
     "SentryAppInstallationForProvider",
     "SentryAppComponent",
     "PlatformExternalIssue",
+    "SentryAppAvatar",
 )
