+++ src/sentry/sentry_apps/web/sentryapp_avatar.py
@@ -1,4 +1,4 @@
-from sentry.models.avatars.sentry_app_avatar import SentryAppAvatar
+from sentry.sentry_apps.models.sentry_app_avatar import SentryAppAvatar
 from sentry.web.frontend.base import AvatarPhotoView, control_silo_view
